NourishNet-Backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.3
Flask-Bcrypt==1.0.1
pymongo==4.5.0
python-dotenv==1.0.0
marshmallow==3.20.1
flask-marshmallow==0.15.0
marshmallow-sqlalchemy==0.29.0
email-validator==2.0.0
bcrypt==4.0.1
python-dateutil==2.8.2
requests==2.31.0
stripe==7.4.0
razorpay==1.3.0
python-socketio==5.8.0
flask-socketio==5.3.6
gunicorn==21.2.0
